UX: remove any mention of 'swap' (#72)
* remove references to 'swap'

* raplace swap with 'transfer' and 'receive'

* spell 'transferring' correctly
