+++ src/components/modals/ConfirmEcashReceiveModal.tsx
@@ -5,7 +5,7 @@ import { useEffect, useState } from 'react';
 import { useSelector } from 'react-redux';
 import { useCashu } from '@/hooks/useCashu';
 import { useToast } from '@/hooks/useToast';
-import ProcessingSwapModal from '../sidebar/ProcessingSwapModal';
+import ProcessingClaimModal from './ProcessingCashuSwap/ProcessingClaim';
 import { useExchangeRate } from '@/hooks/useExchangeRate';
 import { TxStatus, addTransaction } from '@/redux/slices/HistorySlice';
 import { useCashu2 } from '@/hooks/useCashu2';
@@ -280,7 +280,7 @@ const ConfirmEcashReceiveModal = ({ isOpen, token, onClose }: ConfirmEcashReceiv
 
                   <div className='flex flex-col md:flex-row md:justify-center justify-center items-center'>
                      <button onClick={handleSwapToMain} className='mr-3 underline text-lg mb-0'>
-                        {fromActiveMint ? 'Claim' : 'Swap to Main'}
+                        Claim
                      </button>
                      <button
                         className={`underline hover:cursor-pointer text-lg mb-0 ${fromActiveMint ? 'hidden' : ''} ${!supportedUnits.includes('usd') && 'hidden'}`}
@@ -292,7 +292,7 @@ const ConfirmEcashReceiveModal = ({ isOpen, token, onClose }: ConfirmEcashReceiv
                </div>
             </Modal.Body>
          </Modal>
-         <ProcessingSwapModal isSwapping={swapping} />
+         <ProcessingClaimModal isSwapping={swapping} />
       </>
    );
 };
