+++ src/components/modals/ProcessingCashuSwap/ProcessingSwap.tsx
@@ -1,0 +1,17 @@
+import ProcessingModal from './ProcessingModal';
+
+interface ProcessingSwapModalProps {
+   isSwapping: boolean;
+}
+
+const ProcessingSwapModal = ({ isSwapping }: ProcessingSwapModalProps) => {
+   return (
+      <ProcessingModal
+         isSwapping={isSwapping}
+         title='Transferring'
+         prompt='Please wait while we transfer your tokens.'
+      />
+   );
+};
+
+export default ProcessingSwapModal;
